Added separate implementations of each tilemap strategy.
