+++ examples/index.html
@@ -38,7 +38,7 @@
       <a-entity light="type: directional; color: #FFF; intensity: 0.6" position="-0.5 1 1"></a-entity>
 
 
-      <a-entity instanced-tilemap="src: #tilemap-image" position="0 0 -20">
+      <a-entity tilemap-instanced="src: #tilemap-image" position="0 0 -20">
         <a-entity tile="id: 1" geometry="primitive: box; buffer: false" visible="false" scale="1 1 3"></a-entity>
         <a-entity tile="id: 2; isLoaded: true" gltf-model="#milkTruck" visible="false" scale="0.5 0.5 0.5"></a-entity>
         <a-entity tile="id: 3" geometry="primitive: sphere; buffer: false" visible="false"></a-entity>
