+++ src/tilemap-instanced.js
@@ -1,0 +1,238 @@
+import * as THREE from 'three';
+import AFRAME from 'aframe';
+
+import { updateUniforms } from './conversions';
+import { SHADERLIB_MATERIALS, M_TAU_SCALED, Z_AXIS } from './constants';
+
+AFRAME.registerComponent('tilemap-instanced', {
+  schema: {
+    src: { type: 'asset' },
+    tileWidth: { type: 'number', default: 2 },
+    tileHeight: { type: 'number', default: 2 },
+    origin: { type: 'vec2', default: { x: 0.5, y: 0.5 } },
+    debug: { type: 'boolean', default: true },
+  },
+
+  init() {
+    const el = this.el;
+    const tiles = (this.tiles = {});
+
+    // Record all current tile children of this component.
+    for (const child of el.children) {
+      const tile = child.components.tile;
+      if (tile) {
+        tiles[tile.data.id] = {
+          entity: tile,
+          meshes: {},
+          instances: { offsets: [] },
+        };
+      }
+    }
+
+    // TODO: add event handler for new children.
+    // Construct tilemap after a number of pre-processing steps.
+    this.constructTiles().then(() => {
+      this.constructInstances();
+      this.constructMeshes();
+    });
+  },
+
+  // Take all map geometry and add it as meshes to the scene.
+  constructMeshes() {
+    const t0 = performance.now();
+    const tiles = this.tiles;
+
+    for (const tileId in tiles) {
+      const tile = tiles[tileId];
+      const instances = tile.instances;
+      if (instances.offsets.length <= 0) continue;
+
+      // Create instance attributes for all meshes in this tile.
+      const offsetAttribute = new THREE.InstancedBufferAttribute(
+        new Float32Array(instances.offsets),
+        3,
+      );
+
+      // Iterate over each mesh in this tile.
+      for (const uuid in tile.meshes) {
+        const mesh = tile.meshes[uuid];
+        const meshGeometry = mesh.geometry;
+        const meshMaterial = mesh.mesh.material;
+
+        const shader = SHADERLIB_MATERIALS[meshMaterial.type];
+        const uniforms = THREE.UniformsUtils.clone(shader.uniforms);
+        updateUniforms(uniforms, meshMaterial);
+
+        //console.log(uuid);
+        //console.log(meshMaterial);
+        //console.log(uniforms);
+
+        const instanceMaterial = new THREE.ShaderMaterial({
+          uniforms,
+          //vertexShader: shader.vertexShader, // document.getElementById('vertexShader').textContent,
+          vertexShader: document.getElementById('vertexShader').textContent,
+          fragmentShader: shader.fragmentShader,
+          lights: meshMaterial.lights,
+          defines: {
+            USE_MAP: !!meshMaterial.map,
+            USE_ENVMAP: !!meshMaterial.envMap,
+            USE_AOMAP: !!meshMaterial.aoMap,
+            USE_EMISSIVEMAP: !!meshMaterial.emissiveMap,
+            USE_BUMPMAP: !!meshMaterial.bumpMap,
+            USE_NORMALMAP: !!meshMaterial.normalMap,
+            USE_SPECULARMAP: !!meshMaterial.specularMap,
+            USE_ROUGHNESSMAP: !!meshMaterial.roughnessMap,
+            USE_METALNESSMAP: !!meshMaterial.metalnessMap,
+            USE_ALPHAMAP: !!meshMaterial.alphaMap,
+            USE_COLOR: !!meshMaterial.vertexColors,
+            FLAT_SHADED: !!meshMaterial.flatShading,
+            DOUBLE_SIDED: !!meshMaterial.doubleSided,
+            FLIP_SIDED: !meshMaterial.flipSided,
+          },
+        });
+
+        const instanceGeometry = new THREE.InstancedBufferGeometry();
+        instanceGeometry.index = meshGeometry.index;
+        for (const attribute in meshGeometry.attributes) {
+          instanceGeometry.addAttribute(
+            attribute,
+            meshGeometry.getAttribute(attribute),
+          );
+        }
+        instanceGeometry.addAttribute('tilemapOffset', offsetAttribute);
+
+        const instance = new THREE.Mesh(instanceGeometry, instanceMaterial);
+        this.el.object3D.add(instance);
+      }
+    }
+
+    // If the debug flag is set, print timing metrics.
+    if (this.data.debug) {
+      const t1 = performance.now();
+      console.log(`Tile mesh creation took ${(t1 - t0).toFixed(2)} ms.`);
+    }
+  },
+
+  // 1. Get image from this.data.
+  // 2. For each pixel in image.
+  // 3. If the pixel value is in this.tiles.
+  // 4. Add that tile at the corresponding position and rotation.
+  constructInstances() {
+    const t0 = performance.now();
+    const tiles = this.tiles;
+
+    const img = this.data.src;
+    const imgWidth = img.naturalWidth;
+    const imgHeight = img.naturalHeight;
+
+    const tileWidth = this.data.tileWidth;
+    const tileHeight = this.data.tileHeight;
+    const tileOffsetX = -tileWidth * imgWidth * this.data.origin.x;
+    const tileOffsetY = -tileHeight * imgHeight * this.data.origin.y;
+
+    const canvas = document.createElement('canvas');
+    const context = canvas.getContext('2d');
+    context.drawImage(img, 0, 0);
+    const data = context.getImageData(0, 0, imgWidth, imgHeight).data;
+
+    let index = 0;
+    for (let row = 0; row < imgHeight; ++row) {
+      for (let col = 0; col < imgWidth; ++col) {
+        // Extract the pixel components used for the tile rasterization.
+        const [r, g, b, a] = data.slice(index, index + 4);
+        index += 4;
+
+        // Compute the tileId and rotation associated with this tile.
+        const tileId = 256 * r + g;
+
+        // Retrieve the appropriate tile geometry and merge it into place.
+        if (tileId in tiles) {
+          // Determine instance and tile position.
+          const instances = tiles[tileId].instances;
+          const x = tileWidth * col + tileOffsetX;
+          const y = tileHeight * row + tileOffsetY;
+          const theta = M_TAU_SCALED * b;
+
+          // Add this instance's position to the instanced attributes.
+          instances.offsets.push(x, y, theta);
+        }
+      }
+    }
+
+    // If the debug flag is set, print timing metrics.
+    if (this.data.debug) {
+      const t1 = performance.now();
+      console.log(`Tile instancing took ${(t1 - t0).toFixed(2)} ms.`);
+    }
+  },
+
+  constructTiles() {
+    const t0 = performance.now();
+    const tiles = this.tiles;
+    const tileLoadingPromises = [];
+
+    this.el.object3D.updateMatrixWorld();
+    const invMatrixWorld = new THREE.Matrix4().getInverse(
+      this.el.object3D.matrixWorld,
+    );
+
+    for (const tileId in tiles) {
+      const tile = tiles[tileId];
+      const meshes = tile.meshes;
+
+      const tileLoadingPromise = new Promise((resolve, reject) => {
+        const defineTile = () => {
+          tile.entity.el.object3D.traverse(mesh => {
+            if (mesh.type !== 'Mesh') return;
+
+            const geometry =
+              mesh.geometry instanceof THREE.BufferGeometry
+                ? new THREE.BufferGeometry().copy(mesh.geometry)
+                : new THREE.BufferGeometry().fromGeometry(mesh.geometry);
+
+            mesh.updateMatrixWorld();
+            const matrix = new THREE.Matrix4()
+              .copy(invMatrixWorld)
+              .multiply(mesh.matrixWorld);
+            geometry.applyMatrix(matrix);
+
+            meshes[mesh.uuid] = { mesh, geometry };
+          });
+
+          resolve();
+        };
+
+        if (tile.entity.data.isLoaded) {
+          tile.entity.el.addEventListener('model-loaded', e => {
+            // For some reason, there is some additional time for the
+            // transformations in the mesh.matrixWorld to update after the
+            // 'model-loaded' event is emitted.
+            setTimeout(() => {
+              defineTile();
+            }, 100);
+          });
+        } else {
+          defineTile();
+        }
+      });
+
+      tileLoadingPromises.push(tileLoadingPromise);
+    }
+
+    // If the debug flag is set, print timing metrics.
+    if (this.data.debug) {
+      const t1 = performance.now();
+      console.log(`Tile definition took ${(t1 - t0).toFixed(2)} ms.`);
+    }
+
+    return Promise.all(tileLoadingPromises);
+  },
+
+  update(oldData) {
+    // TODO: Regenerate mesh if these properties change.
+  },
+
+  remove() {
+    // Do nothing.
+  },
+});
